#334: Move test helpers to helpers folder
