+++ CourageScores/ClientApp/src/components/division_fixtures/FilterFixtures.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp} from "../../tests/helpers";
+import {cleanUp, renderApp} from "../../helpers/tests";
 import {renderDate} from "../../helpers/rendering";
 import React from "react";
 import {FilterFixtures} from "./FilterFixtures";
