+++ CourageScores/ClientApp/src/components/division_fixtures/sayg/MatchAverage.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp, doClick} from "../../../tests/helpers";
+import {cleanUp, renderApp, doClick} from "../../../helpers/tests";
 import React from "react";
 import {MatchAverage} from "./MatchAverage";
