+++ CourageScores/ClientApp/src/components/division_fixtures/sayg/MatchStatistics.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp} from "../../../tests/helpers";
+import {cleanUp, renderApp} from "../../../helpers/tests";
 import React from "react";
 import {MatchStatistics} from "./MatchStatistics";
