+++ CourageScores/ClientApp/src/components/division_fixtures/scores/MatchPlayerSelection.test.js
@@ -1,7 +1,7 @@
 // noinspection JSUnresolvedFunction
 
 import React from "react";
-import {cleanUp, renderApp, doClick, doChange} from "../../../tests/helpers";
+import {cleanUp, renderApp, doClick, doChange} from "../../../helpers/tests";
 import {MatchPlayerSelection, NEW_PLAYER} from "./MatchPlayerSelection";
 import {createTemporaryId} from "../../../helpers/projection";
 import {LeagueFixtureContainer} from "../LeagueFixtureContainer";
