+++ CourageScores/ClientApp/src/components/division_fixtures/scores/MergeHiCheckAnd180s.test.js
@@ -1,7 +1,7 @@
 // noinspection JSUnresolvedFunction
 
 import React from "react";
-import {cleanUp, renderApp, doClick, findButton} from "../../../tests/helpers";
+import {cleanUp, renderApp, doClick, findButton} from "../../../helpers/tests";
 import {createTemporaryId} from "../../../helpers/projection";
 import {MergeHiCheckAnd180s} from "./MergeHiCheckAnd180s";
