+++ CourageScores/ClientApp/src/components/division_fixtures/scores/MultiPlayerSelection.test.js
@@ -1,7 +1,7 @@
 // noinspection JSUnresolvedFunction
 
 import React from "react";
-import {cleanUp, doClick, renderApp, doChange} from "../../../tests/helpers";
+import {cleanUp, doClick, renderApp, doChange} from "../../../helpers/tests";
 import {MultiPlayerSelection} from "./MultiPlayerSelection";
 import {createTemporaryId} from "../../../helpers/projection";
