+++ CourageScores/ClientApp/src/components/division_fixtures/scores/ScoreCardHeading.test.js
@@ -1,7 +1,7 @@
 // noinspection JSUnresolvedFunction
 
 import React from "react";
-import {cleanUp, renderApp, doClick} from "../../../tests/helpers";
+import {cleanUp, renderApp, doClick} from "../../../helpers/tests";
 import {ScoreCardHeading} from "./ScoreCardHeading";
 import {createTemporaryId} from "../../../helpers/projection";
