+++ CourageScores/ClientApp/src/components/division_fixtures/tournaments/Tournament.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp, doClick, doChange, findButton} from "../../../tests/helpers";
+import {cleanUp, renderApp, doClick, doChange, findButton} from "../../../helpers/tests";
 import React from "react";
 import {Tournament} from "./Tournament";
 import {toMap, any} from "../../../helpers/collections";
