+++ CourageScores/ClientApp/src/components/division_fixtures/tournaments/TournamentSheet.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp} from "../../../tests/helpers";
+import {cleanUp, renderApp} from "../../../helpers/tests";
 import React from "react";
 import {createTemporaryId, repeat} from "../../../helpers/projection";
 import {TournamentSheet} from "./TournamentSheet";
