+++ CourageScores/ClientApp/src/components/division_players/PlayerOverview.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp} from "../../tests/helpers";
+import {cleanUp, renderApp} from "../../helpers/tests";
 import React from "react";
 import {DivisionDataContainer} from "../DivisionDataContainer";
 import {createTemporaryId} from "../../helpers/projection";
