+++ CourageScores/ClientApp/src/components/division_teams/DivisionTeam.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp, doClick, doChange, findButton} from "../../tests/helpers";
+import {cleanUp, renderApp, doClick, doChange, findButton} from "../../helpers/tests";
 import React from "react";
 import {createTemporaryId} from "../../helpers/projection";
 import {DivisionTeam} from "./DivisionTeam";
