+++ CourageScores/ClientApp/src/components/layout/Layout.test.js
@@ -1,6 +1,6 @@
 // noinspection JSUnresolvedFunction
 
-import {cleanUp, renderApp} from "../../tests/helpers";
+import {cleanUp, renderApp} from "../../helpers/tests";
 import React from "react";
 import {Layout} from "./Layout";
